requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.15.0=pyhd3eb1b0_0
addict=2.4.0=pypi_0
aiohttp=3.8.1=py37h2bbff1b_0
aiosignal=1.2.0=pyhd3eb1b0_0
alabaster=0.7.12=pypi_0
astor=0.8.1=py37haa95532_0
astroid=2.9.3=pypi_0
async-timeout=4.0.1=pyhd3eb1b0_0
asynctest=0.13.0=py_0
atomicwrites=1.4.0=py_0
attrs=19.3.0=pypi_0
autopep8=1.5.7=pyhd3eb1b0_0
babel=2.9.1=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
backports-entry-points-selectable=1.1.0=pypi_0
bayesian-optimization=1.2.0=pypi_0
bce-python-sdk=0.8.62=pypi_0
blas=1.0=mkl
blinker=1.4=py37haa95532_0
blosc=1.21.0=h19a0ad4_0
brotli=1.0.9=ha925a31_2
brotlipy=0.7.0=py37h2bbff1b_1003
bzip2=1.0.8=he774522_0
ca-certificates=2022.10.11=haa95532_0
cached-property=1.5.2=py_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2022.9.24=py37haa95532_0
cffi=1.15.0=py37h2bbff1b_1
cfgv=3.3.1=pypi_0
cfitsio=3.470=he774522_6
charls=2.2.0=h6c2663c_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py37haa95532_0
cloudpickle=2.0.0=pyhd3eb1b0_0
cma=3.1.0=pypi_0
colorama=0.4.4=pyhd3eb1b0_0
coverage=6.3.2=py37h2bbff1b_0
cpuonly=1.0=0
cryptography=3.4.8=py37h71e12ea_0
cycler=0.11.0=pyhd3eb1b0_0
cytoolz=0.11.0=py37he774522_0
dask-core=2021.10.0=pyhd3eb1b0_0
dataclasses=0.8=pyh6d0b6a4_7
decorator=5.1.1=pyhd3eb1b0_0
decord=0.6.0=pypi_0
distance=0.1.3=pypi_0
distlib=0.3.3=pypi_0
dnspython=2.1.0=pypi_0
docopt=0.6.2=pypi_0
docutils=0.17.1=pypi_0
entrypoints=0.3=py37_0
fett=0.3.2=pypi_0
filelock=3.3.1=pypi_0
flake8=4.0.1=pypi_0
flask=2.0.2=pypi_0
flask-babel=2.0.0=pypi_0
freetype=2.10.4=hd328e21_0
frozenlist=1.2.0=py37h2bbff1b_0
fsspec=2022.1.0=pyhd3eb1b0_0
future=0.18.2=pypi_0
gast=0.3.3=py_0
giflib=5.2.1=h62dcd97_0
google-auth=1.33.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.1=py_2
gputil=1.4.0=pypi_0
grpcio=1.42.0=py37hc60d5dd_0
h5py=3.2.1=py37h3de5c98_0
hdf5=1.10.6=h7ebc959_0
huggingface-hub=0.0.19=pypi_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
identify=2.3.1=pypi_0
idna=3.2=pypi_0
imagecodecs=2021.8.26=py37ha1f97ea_0
imageio=2.9.0=pyhd3eb1b0_0
imagesize=1.3.0=pypi_0
importlib-metadata=1.7.0=pypi_0
importlib_metadata=4.8.2=hd3eb1b0_0
iniconfig=1.1.1=pyhd3eb1b0_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=5.3.0=py37h5ca1d4c_0
ipython=7.31.1=py37haa95532_0
isort=5.9.3=pypi_0
itsdangerous=2.0.1=pypi_0
jedi=0.18.1=py37haa95532_1
jinja2=3.0.2=pypi_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9d=h2bbff1b_0
jsonpatch=1.32=pypi_0
jsonpointer=2.1=pypi_0
jsonschema=3.2.0=pypi_0
jupyter_client=7.1.2=pyhd3eb1b0_0
jupyter_core=4.9.1=py37haa95532_0
jupytext=1.5.0=pypi_0
kiwisolver=1.3.2=py37hd77b12b_0
lazy-object-proxy=1.6.0=pypi_0
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libaec=1.0.4=h33f27b4_1
libdeflate=1.8=h2bbff1b_5
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.19.1=h23ce68f_0
libtiff=4.2.0=hd0e1b90_0
libwebp=1.2.2=h2bbff1b_0
libzopfli=1.0.3=ha925a31_0
lmdb=1.2.1=pypi_0
locket=0.2.1=py37haa95532_1
lz4-c=1.9.3=h2bbff1b_1
markdown=3.3.4=py37haa95532_0
markupsafe=2.0.1=pypi_0
matplotlib=3.3.2=py37hc8dfbb8_1
matplotlib-base=3.3.2=py37h64d6e1c_1
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mccabe=0.6.1=pypi_0
medpy=0.4.0=pypi_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py37h2bbff1b_0
mkl_fft=1.3.1=py37h277e83a_0
mkl_random=1.2.2=py37hf11a4ad_0
mmcv=1.3.16=pypi_0
monai=1.1.0=pypi_0
more-itertools=8.12.0=pyhd3eb1b0_0
multidict=5.1.0=py37h2bbff1b_2
mypy=0.910=pypi_0
mypy-extensions=0.4.3=pypi_0
nbformat=5.0.7=pypi_0
nest-asyncio=1.5.1=pyhd3eb1b0_0
networkx=2.6.3=pyhd3eb1b0_0
nevergrad=0.4.3.post8=pypi_0
nibabel=3.1.1=pypi_0
ninja=1.10.2=py37h559b2a2_3
nodeenv=1.6.0=pypi_0
numpy=1.21.5=py37ha4e8547_0
numpy-base=1.21.5=py37hc2deb75_0
oauthlib=3.2.0=pyhd3eb1b0_0
onnx=1.10.1=pypi_0
opencv-python=4.5.3.56=pypi_0
openjpeg=2.4.0=h4fc8c34_0
openssl=1.1.1q=h2bbff1b_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=20.4=pypi_0
paddlepaddle=2.3.1=py37_cpu_windows
pandas=1.2.4=py37hf11a4ad_0
parso=0.8.3=pyhd3eb1b0_0
partd=1.2.0=pyhd3eb1b0_0
pathspec=0.9.0=pypi_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.0.1=py37hdc2b20a_0
pip=21.3.1=pypi_0
platformdirs=2.4.0=pypi_0
pluggy=0.13.1=py37haa95532_0
pre-commit=2.15.0=pypi_0
progressbar=2.5=pypi_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.19.1=py37hd77b12b_0
psutil=5.8.0=pypi_0
py=1.11.0=pyhd3eb1b0_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycodestyle=2.8.0=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pycryptodome=3.11.0=pypi_0
pydantic=1.8.2=pypi_0
pyflakes=2.4.0=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
pyjwt=2.1.0=py37haa95532_0
pylint=2.12.0=pypi_0
pymongo=3.12.1=pypi_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py37hd77b12b_6
pyreadline=2.1=py37_1
pyrsistent=0.16.0=pypi_0
pysocks=1.7.1=py37_1
pytest=6.2.4=py37haa95532_2
pytest-cov=3.0.0=pypi_0
python=3.7.7=h81c818b_4
python-dateutil=2.8.2=pyhd3eb1b0_0
python-jsonrpc-server=0.3.4=pypi_0
python_abi=3.7=1_cp37m
pytz=2021.3=pyhd3eb1b0_0
pywavelets=1.1.1=py37he774522_2
pywin32=302=py37h827c3e9_1
pyyaml=5.3.1=pypi_0
pyzmq=22.3.0=py37hd77b12b_2
qt=5.9.7=vc14h73c81de_0
regex=2021.10.23=pypi_0
reprod-log=1.0.0=pypi_0
requests=2.26.0=pypi_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
sacremoses=0.0.46=pypi_0
scikit-image=0.18.1=py37hf11a4ad_0
scikit-learn=0.24.2=py37hf11a4ad_1
scipy=1.7.1=pypi_0
seaborn=0.11.2=pyhd3eb1b0_0
setuptools=58.0.4=py37haa95532_0
shellcheck-py=0.7.2.1=pypi_0
simpleitk=2.1.0=pypi_0
sip=4.19.13=py37hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.8=h33f27b4_0
snooty-lextudio=1.11.5=pypi_0
snowballstemmer=2.2.0=pypi_0
sphinx=4.3.0=pypi_0
sphinx-rtd-theme=1.0.0=pypi_0
sphinxcontrib-applehelp=1.0.2=pypi_0
sphinxcontrib-devhelp=1.0.2=pypi_0
sphinxcontrib-htmlhelp=2.0.0=pypi_0
sphinxcontrib-jsmath=1.0.1=pypi_0
sphinxcontrib-qthelp=1.0.3=pypi_0
sphinxcontrib-serializinghtml=1.1.5=pypi_0
sqlite=3.37.2=h2bbff1b_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2021.7.2=pyhd3eb1b0_2
tk=8.6.11=h2bbff1b_0
tokenizers=0.10.3=pypi_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.11.2=pyhd3eb1b0_0
tornado=6.1=py37h2bbff1b_0
tqdm=4.62.0=pyhd3eb1b0_1
traitlets=5.1.1=pyhd3eb1b0_0
transformers=4.11.3=pypi_0
treelib=1.6.1=pypi_0
typed-ast=1.4.3=pypi_0
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.8=pyhd3eb1b0_0
vc=14.2=h21ff451_1
virtualenv=20.9.0=pypi_0
visdom=0.1.8.9=pypi_0
visualdl=2.2.1=pypi_0
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=1.0.2=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
websocket-client=1.1.1=pypi_0
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py37haa95532_0
wincertstore=0.2=py37haa95532_2
wrapt=1.13.2=pypi_0
x2paddle=1.3.1=pypi_0
xlrd=2.0.1=pyhd3eb1b0_0
xz=5.2.5=h62dcd97_0
yaml=0.2.5=he774522_0
yapf=0.31.0=pypi_0
yarl=1.6.3=py37h2bbff1b_0
zfp=0.5.5=hd77b12b_6
zipp=3.1.0=pypi_0
zlib=1.2.11=h8cc25b3_4
zstd=1.4.9=h19a0ad4_0
